refactor(knockdog): 상수 분리 및 constants 디렉토리 단수로 변경 (constant)
