+++ apps/knockdog/src/entities/social-user/model/constant/provider.ts
@@ -1,7 +1,13 @@
+const SOCIAL_PROVIDER = {
+  KAKAO: 'KAKAO',
+  GOOGLE: 'GOOGLE',
+  APPLE: 'APPLE',
+} as const;
+
 const SOCIAL_PROVIDER_KO = {
   KAKAO: '카카오',
   GOOGLE: '구글',
   APPLE: '애플',
 } as const;
 
-export { SOCIAL_PROVIDER_KO };
+export { SOCIAL_PROVIDER, SOCIAL_PROVIDER_KO };
