+++ apps/knockdog/src/entities/social-user/model/socialUser.ts
@@ -1,8 +1,4 @@
-const SOCIAL_PROVIDER = {
-  KAKAO: 'KAKAO',
-  GOOGLE: 'GOOGLE',
-  APPLE: 'APPLE',
-} as const;
+import { SOCIAL_PROVIDER } from './constant/provider';
 
 type SocialProvider = (typeof SOCIAL_PROVIDER)[keyof typeof SOCIAL_PROVIDER];
