--- apps/knockdog/src/entities/user/model/constants/user.ts
@@ -1,8 +1,0 @@
-/** 주소 타입 한글 */
-const USER_ADDRESS_TYPE_KR = {
-  HOME: '집',
-  WORK: '회사',
-  OTHER: '기타',
-} as const;
-
-export { USER_ADDRESS_TYPE_KR };
