+++ apps/knockdog/src/entities/user/model/user.ts
@@ -1,3 +1,5 @@
+import { USER_STATUS, USER_ADDRESS_TYPE } from './constant/user';
+
 /** 유저 */
 interface User {
   status: UserStatus;
@@ -8,13 +10,6 @@ interface User {
 }
 
 /** 유저 상태 */
-const USER_STATUS = {
-  /** 활동중인 유저 */
-  ACTIVE: 'ACTIVE',
-  /** 탈퇴한 유저 */
-  WITHDRAWN: 'WITHDRAWN',
-} as const;
-
 type UserStatus = (typeof USER_STATUS)[keyof typeof USER_STATUS];
 
 /** 주소 */
@@ -30,15 +25,6 @@ interface UserAddress {
 }
 
 /** 주소 타입 */
-const USER_ADDRESS_TYPE = {
-  /** 자택 */
-  HOME: 'HOME',
-  /** 직장 */
-  WORK: 'WORK',
-  /** 기타 */
-  OTHER: 'OTHER',
-} as const;
-
 type UserAddressType = (typeof USER_ADDRESS_TYPE)[keyof typeof USER_ADDRESS_TYPE];
 
 export { USER_STATUS, USER_ADDRESS_TYPE, type User, type UserStatus, type UserAddress, type UserAddressType };
